--- a/talentmap_api/language/models.py
+++ b/talentmap_api/language/models.py
@@ -3,8 +3,10 @@ from djchoices import DjangoChoices, ChoiceItem
 
 import logging
 
+from talentmap_api.common.models import StaticRepresentationModel
 
-class Language(models.Model):
+
+class Language(StaticRepresentationModel):
     '''
     The language model represents an individual language, such as English, French, et. al.
     It is typically linked together with a language proficiency to create a qualification,
@@ -24,7 +26,7 @@ class Language(models.Model):
         ordering = ["code"]
 
 
-class Proficiency(models.Model):
+class Proficiency(StaticRepresentationModel):
     '''
     The language proficiency represents a positions linguistic proficiency requirement,
     or the linguistic proficiency of an individual. These are typically not used as
@@ -52,7 +54,7 @@ class Proficiency(models.Model):
         ordering = ["code"]
 
 
-class Qualification(models.Model):
+class Qualification(StaticRepresentationModel):
     '''
     The language qualification is defined by a combination of language proficiencies
     and a specific language. For example, German 2/2+, where the first numeral denotes
@@ -97,7 +99,7 @@ class Qualification(models.Model):
         unique_together = (('language', 'reading_proficiency', 'spoken_proficiency'))
 
 
-class Waiver(models.Model):
+class Waiver(StaticRepresentationModel):
     '''
     The waiver model represents a language requirement waiver for a particular position
     assignment, along with the status