… talentmap_api/bidding/migrations/0016_auto_20171205_1627.py
… talentmap_api/language/migrations/0010_auto_20171205_1627.py
… talentmap_api/messaging/migrations/0006_auto_20171205_1627.py
… talentmap_api/organization/migrations/0018_auto_20171205_1627.py
… talentmap_api/position/migrations/0026_auto_20171205_1627.py
… talentmap_api/position/migrations/0027_auto_20171205_1813.py
… talentmap_api/position/tests/test_position_endpoints.py
… talentmap_api/position/tests/test_position_full_text_search.py
… talentmap_api/user_profile/migrations/0019_auto_20171205_1627.py
… talentmap_api/user_profile/models.py
… talentmap_api/user_profile/serializers.py