- Add abstract model `StaticRepresentationModel` to provide functionality where an object's string representation is stored on the database. This eases the burden on the server during serialization of objects when calling `to_representation()`
- Switch XML loader to individually call `.save()` to work properly with StaticRepresentationModel save override
- Fix problem with `is_available_in_current_bidcycle` filter where it was throwing away prefetch directives due to re-acquiring queryset from bidcycle - moved to a subquery and sped it up
- Tweak prefetched serializer to (1) be more readable and (2) use nested serializers for continuing prefetching, reducing the amount of sql queries required for most requests
- Also snuck in an update to demo environment creation to populate some bids per @mjoyce91 request
- Moved `language_qualifications` to `languages` on `Position` model to fix a problem where prefetching didn't want to prefetch languages. @mjoyce91 this will probably require a front end change as the field name updated.

STATS:
dev demo box deployment: `/api/v1/position/`  370 queries, 360 ms
this branch: `/api/v1/position/` 9 queries, 108 ms

dev demo box deployment: `/api/v1/position/?is_available_in_current_bidcycle=true` 1769 queries, ~1600 ms
this branch: `/api/v1/position/?is_available_in_current_bidcycle=true` 10 queries, 156 ms