@jseppi 2017-12-05
Also I'm fine with the extremely slight coverage drop (it's not a required check anyway). CodeClimate doesn't appear to have a way to specify an acceptable drop percent -- it's just a boolean of "did the coverage decrease at all?"