@afrasier 2017-12-05
Yeah, that article is basically the problem +  a specific solution, I've just gone through and generalized it a bit and made it a bit more automatic.

If you're ok with me fixing the left-over comment in a later PR, you can admin-merge this in.