@afrasier 2017-12-05
Nevermind, I could do it :)